requirements.txt
requests==2.32.3
PyNaCl==1.5.0
python-dotenv==1.0.1